+++ CHANGELOG.md
@@ -6,7 +6,8 @@ This changelog track changes to the qoqo-quest project starting at version 0.1.0
 
 ## 0.14.0
 
-* Updated to qoqo 1.12.0 and pyo3 0.21
+* Updated to qoqo 1.12.0 and pyo3 0.21.
+* Clearer error message for qubits index out of range.
 
 ## 0.13.1